+++ lib/flycal_cli/slot_formatter.rb
@@ -1,8 +1,19 @@
 # frozen_string_literal: true
 
+require "cgi"
+
 module FlycalCli
   class SlotFormatter
     class << self
+      def format_header(in_value:, duration:, calendars:)
+        lines = []
+        lines << Locale.t("slots.header", in: in_value, duration: duration)
+        calendars.each do |cal|
+          lines << calendar_link(cal[:name], cal[:id])
+        end
+        lines.join("\n")
+      end
+
       def format_output(slots_by_day)
         lines = []
 
@@ -18,6 +29,12 @@ module FlycalCli
       end
 
       private
+
+      def calendar_link(name, calendar_id)
+        url = "https://calendar.google.com/calendar/r?cid=#{CGI.escape(calendar_id.to_s)}"
+        # OSC 8 hyperlink (clickable in modern terminals)
+        "\e]8;;#{url}\a#{name}\e]8;;\a"
+      end
 
       def day_header(date)
         "#{Locale.day_name(date)} #{date.day}/#{date.month}"